Refactor feature generator imports and update documentation structure
