+++ docs/api/features.md
@@ -2,54 +2,54 @@
 
 ## Base Class
 
-::: optimhc.feature_generator.base_feature_generator
+::: optimhc.feature.base_feature_generator
     options:
       members: true
 
 ## Basic
 
-::: optimhc.feature_generator.basic
+::: optimhc.feature.basic
     options:
       members: true
 
 ## Spectral Similarity
 
-::: optimhc.feature_generator.spectral_similarity
+::: optimhc.feature.spectral_similarity
     options:
       members: true
 
 ## DeepLC
 
-::: optimhc.feature_generator.DeepLC
+::: optimhc.feature.DeepLC
     options:
       members: true
 
 ## Overlapping Peptide
 
-::: optimhc.feature_generator.overlapping_peptide
+::: optimhc.feature.overlapping_peptide
     options:
       members: true
 
 ## PWM
 
-::: optimhc.feature_generator.PWM
+::: optimhc.feature.PWM
     options:
       members: true
 
 ## MHCflurry
 
-::: optimhc.feature_generator.mhcflurry
+::: optimhc.feature.mhcflurry
     options:
       members: true
 
 ## NetMHCpan
 
-::: optimhc.feature_generator.netMHCpan
+::: optimhc.feature.netMHCpan
     options:
       members: true
 
 ## NetMHCIIpan
 
-::: optimhc.feature_generator.netMHCIIpan
+::: optimhc.feature.netMHCIIpan
     options:
       members: true
