--- optimhc/feature_generator/base_feature_generator.py
@@ -1,35 +1,0 @@
-# feature_generators/base_feature_generator.py
-
-from abc import ABC, abstractmethod
-from typing import List
-
-import pandas as pd
-
-
-class BaseFeatureGenerator(ABC):
-    """
-    Abstract base class for all feature generators in the rescoring pipeline.
-    """
-
-    @property
-    @abstractmethod
-    def feature_columns(self) -> List[str]:
-        """
-        Returns a list of feature names generated by the feature generator.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def id_column(self) -> List[str]:
-        """
-        Returns the column or columns used as key or keys to merge features with PSMs.
-        """
-        pass
-
-    @abstractmethod
-    def generate_features(self) -> pd.DataFrame:
-        """
-        Generates features.
-        """
-        pass
